+++ src/visitors/reader.js
@@ -1,7 +1,7 @@
-/* eslint-disable max-len, no-param-reassign */
-import { strideof, createMask, createVariable } from './utils';
+/* eslint-disable max-len, no-new-func */
+import { strideof, createMask, createVariable } from '../utils';
 
-export const generateReaderVisitor = {
+export const readerVisitor = Object.freeze({
   Bool(_, stackDepth) {
     const resultVar = createVariable('result', stackDepth);
     const byteOffsetVar = createVariable('byteOffset', stackDepth);
@@ -73,12 +73,19 @@ export const generateReaderVisitor = {
     return `${resultVar} = dataView.getFloat64(${byteOffsetVar}, ${littleEndian});`;
   },
 
+  Number({ littleEndian, kind }, stackDepth) {
+    const resultVar = createVariable('result', stackDepth);
+    const byteOffsetVar = createVariable('byteOffset', stackDepth);
+
+    return `${resultVar} = dataView.get${kind}(${byteOffsetVar}, ${littleEndian});`;
+  },
+
   String({ byteLength, encoding }, stackDepth) {
     const resultVar = createVariable('result', stackDepth);
     const arrayVar = createVariable('array', stackDepth);
     const indexVar = createVariable('i', stackDepth);
-
     const byteOffsetVar = createVariable('byteOffset', stackDepth);
+
     return `
       var ${arrayVar} = new Uint8Array(dataView.buffer, ${byteOffsetVar}, ${byteLength});
       var ${indexVar} = ${arrayVar}.indexOf(0);
@@ -105,7 +112,7 @@ export const generateReaderVisitor = {
       var ${innerByteOffsetVar} = ${byteOffsetVar};
       for (var ${indexVar} = 0; ${indexVar} < ${length}; ++${indexVar}) {
         var ${innerResultVar} = ${resultVar}[${indexVar}];
-        ${generateReaderVisitor[element.tag](element, innerStackDepth)}
+        ${readerVisitor[element.tag](element, innerStackDepth)}
         ${resultVar}[${indexVar}] = ${innerResultVar};
         ${innerByteOffsetVar} += ${stride};
       }
@@ -126,7 +133,7 @@ export const generateReaderVisitor = {
       ${members.map(({ element, byteOffset }, i) => `
         var ${innerResultVar} = ${resultVar}[${i}];
         var ${innerByteOffsetVar} = ${byteOffsetVar} + ${byteOffset};
-        ${generateReaderVisitor[element.tag](element, innerStackDepth)}
+        ${readerVisitor[element.tag](element, innerStackDepth)}
         ${resultVar}[${i}] = ${innerResultVar};
       `).join('\n')}
     `;
@@ -147,7 +154,7 @@ export const generateReaderVisitor = {
       ${members.map(({ name, element, byteOffset }) => `
         var ${innerResultVar} = ${resultVar}[${JSON.stringify(name)}];
         var ${innerByteOffsetVar} = ${byteOffsetVar} + ${byteOffset};
-        ${generateReaderVisitor[element.tag](element, innerStackDepth)}
+        ${readerVisitor[element.tag](element, innerStackDepth)}
         ${resultVar}[${JSON.stringify(name)}] = ${innerResultVar};
       `).join('\n')}
     `;
@@ -167,162 +174,22 @@ export const generateReaderVisitor = {
       }
       var ${innerByteOffsetVar} = ${byteOffsetVar};
       var ${innerResultVar};
-      ${generateReaderVisitor[element.tag](element, innerStackDepth)}
+      ${readerVisitor[element.tag](element, innerStackDepth)}
       ${members.map(({ name, bits }) => `
         ${resultVar}[${JSON.stringify(name)}] = ${innerResultVar} & ${createMask(bits)};
         ${innerResultVar} >>>= ${bits};
       `).join('\n')}
     `;
   },
-};
+});
 
-export const generateWriterVisitor = {
-  Bool(_, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setUint8(${byteOffsetVar}, ${dataVar});`;
-  },
+export function createReader(type) {
+  const resultVar = createVariable('result');
+  const byteOffsetVar = createVariable('byteOffset');
 
-  Int8(_, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setInt8(${byteOffsetVar}, ${dataVar});`;
-  },
-
-  UInt8(_, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setUint8(${byteOffsetVar}, ${dataVar});`;
-  },
-
-  Int16({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setInt16(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  UInt16({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setUint16(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  Int32({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setInt32(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  UInt32({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setUint32(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  UInt64({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    const lowOffset = littleEndian ? 0 : 4;
-    const highOffset = littleEndian ? 4 : 0;
-    return `
-      dataView.setUint32(${byteOffsetVar} + ${lowOffset}, ${dataVar}.lo, ${littleEndian});
-      dataView.setUint32(${byteOffsetVar} + ${highOffset}, ${dataVar}.hi, ${littleEndian});
-    `;
-  },
-
-  Float32({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setFloat32(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  Float64({ littleEndian }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    return `dataView.setFloat64(${byteOffsetVar}, ${dataVar}, ${littleEndian});`;
-  },
-
-  String({ byteLength, encoding }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const indexVar = createVariable('i', stackDepth);
-    const lengthVar = createVariable('length', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-
-    return `
-      ${dataVar} = new String(${dataVar});
-      for (var ${indexVar} = 0, ${lengthVar} = ${dataVar}.length; ${indexVar} < ${lengthVar}; ++${indexVar}) {
-        dataView.setUint8(${byteOffsetVar} + ${indexVar}, ${dataVar}.charCodeAt(${indexVar}));
-      }
-    `;
-  },
-
-  Array({ byteAlignment, length, element, pack }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    const indexVar = createVariable('i', stackDepth);
-    const innerStackDepth = stackDepth + 1;
-    const innerDataVar = createVariable('data', innerStackDepth);
-    const innerByteOffsetVar = createVariable('byteOffset', innerStackDepth);
-    const stride = strideof(element, byteAlignment);
-
-    return `
-      var ${innerByteOffsetVar} = ${byteOffsetVar};
-      for (var ${indexVar} = 0; ${indexVar} < ${length}; ++${indexVar}) {
-        var ${innerDataVar} = ${dataVar}[${indexVar}];
-        ${generateWriterVisitor[element.tag](element, innerStackDepth)}
-        ${innerByteOffsetVar} += ${stride};
-      }
-    `;
-  },
-
-  Tuple({ members }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    const innerStackDepth = stackDepth + 1;
-    const innerDataVar = createVariable('data', innerStackDepth);
-    const innerByteOffsetVar = createVariable('byteOffset', innerStackDepth);
-
-    return `
-      ${members.map(({ element, byteOffset }, i) => `
-        var ${innerDataVar} = ${dataVar}[${i}];
-        var ${innerByteOffsetVar} = ${byteOffsetVar} + ${byteOffset};
-        ${generateWriterVisitor[element.tag](element, innerStackDepth)}
-      `).join('\n')}
-    `;
-  },
-
-  Struct({ members }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    const innerStackDepth = stackDepth + 1;
-    const innerDataVar = createVariable('data', innerStackDepth);
-    const innerByteOffsetVar = createVariable('byteOffset', innerStackDepth);
-
-    return `
-      ${members.map(({ name, element, byteOffset }) => `
-        var ${innerDataVar} = ${dataVar}[${JSON.stringify(name)}];
-        var ${innerByteOffsetVar} = ${byteOffsetVar} + ${byteOffset};
-        ${generateWriterVisitor[element.tag](element, innerStackDepth)}
-      `).join('\n')}
-    `;
-  },
-
-  Bitfield({ element, members }, stackDepth) {
-    const dataVar = createVariable('data', stackDepth);
-    const byteOffsetVar = createVariable('byteOffset', stackDepth);
-    const innerStackDepth = stackDepth + 1;
-    const innerDataVar = createVariable('data', innerStackDepth);
-    const innerByteOffsetVar = createVariable('byteOffset', innerStackDepth);
-
-    return `
-      var ${innerByteOffsetVar} = ${byteOffsetVar};
-      var ${innerDataVar} = 0;
-      ${members.slice().reverse().map(({ name, bits }) => `
-        ${innerDataVar} <<= ${bits};
-        ${innerDataVar} |= ${dataVar}[${JSON.stringify(name)}] & ${createMask(bits)};
-      `).join('\n')}
-      ${generateWriterVisitor[element.tag](element, innerStackDepth)}
-    `;
-  },
-};
-
+  return new Function('dataView', byteOffsetVar, resultVar, `
+    "use strict";
+    ${readerVisitor[type.tag](type, 0)}
+    return ${resultVar};
+  `);
+}
